requirements.txt
absl-py==1.1.0
albumentations==1.2.0
argon2-cffi
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
azure-core
azure-storage-blob==12.13.1
azureml-sdk[notebooks]
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blis==0.7.7
cachetools==5.2.0
catalogue
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
contextlib2==21.6.0
cryptography==37.0.4
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastai==2.7.3
fastcore==1.4.5
fastdownload==0.0.6
fastjsonschema==2.15.3
fastprogress==1.0.2
fonttools==4.33.3
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.19.3
importlib-metadata
importlib-resources==5.8.0
ipdb==0.13.9
ipykernel==6.15.1
ipython
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema
jupyter-client
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaggle==1.5.12
kiwisolver
langcodes==3.3.0
llvmlite==0.39.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
ml-collections==0.1.1
msrest==0.6.21
murmurhash==1.0.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx
notebook==6.4.12
numba==0.56.0
numpy
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
qudida
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.3
scikit-learn
scipy
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.15.0
text-unidecode==1.3
thinc==8.0.17
threadpoolctl==3.1.0
tifffile
timm==0.5.4
tinycss2==1.1.1
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.1
typing_extensions
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
zipp==3.8.0
